% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_methylome.R
\name{annotate_methylome}
\alias{annotate_methylome}
\title{annotate_methylome()
This function annotates the methylated sites with different genomic features (e.g. genes, exon, promoter, etc.)}
\usage{
annotate_methylome(mrobj, genome_info, save_output_data = TRUE,
  outputdata = "annotate_methylome_data.Rda")
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object that is generated by mcalls2mkobj()}

\item{genome_info}{A list of GRanges objects that contains genome infomation.}

\item{save_output_data}{If specified other than the default "TRUE", then results
are saved in Rda file that is specified by @param outputdata; otherwise either no Rda file is generated.}

\item{outputdata}{If specified, then the result is saved in the specified file name; otherwise the result
is saved in '$wd/annotate_methylome_data.Rda'}
}
\value{
A table ?that stores the methylation information and annotate_methylome
}
\description{
annotate_methylome()
This function annotates the methylated sites with different genomic features (e.g. genes, exon, promoter, etc.)
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  ginfo <- get_genome_annotation(myfiles$parameters)
  annotate_methylome(AmHE, ginfo, save_output_data=TRUE, outputdata='AmHE_annotate_methylome.Rda')

}
