% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_cal_methyl_level_in_gene.R
\name{cal_methyl_level_in_gene}
\alias{cal_methyl_level_in_gene}
\title{annotate_methylome()
This function calculate the number of methylated sites and average methylation level for each gene}
\usage{
cal_methyl_level_in_gene(mrobj, genome_info,
  outputdata = "methyl_level_in_gene.txt")
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object that is generated by mcalls2mkobj()}

\item{genome_info}{A list of GRanges objects that contains genome infomation.}

\item{outputdata}{If specified, then the result is saved in the specified file name; otherwise the result
is saved in 'methyl_level_in_gene'}
}
\value{
A table that stores the information of methylation level in each gene
}
\description{
annotate_methylome()
This function calculate the number of methylated sites and average methylation level for each gene
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  ginfo <- get_genome_annotation(myfiles$parameters)
  cal_methyl_level_in_gene(AmHE, ginfo, outputdata='AmHE_methyl_level_in_gene.txt')

}
