% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg_deps.R
\name{check_pkg_deps}
\alias{check_pkg_deps}
\title{check_pkg_deps()
This function import the libraries that are required}
\usage{
check_pkg_deps()
}
\value{
If required packages are not installed, return a message. If required packages are installed, return nothing
}
\description{
check_pkg_deps()
This function import the libraries that are required
}
\examples{
  check_pkg_deps()

}
