% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpSamples.R
\name{cmpSamples}
\alias{cmpSamples}
\title{cmpSamples()
This function will methylKit::unite() input samples and compare
them using methylKit::getCorrelation() and methylKit::PCASamples.}
\usage{
cmpSamples(mrobj, destrand = FALSE, plot2file = FALSE)
}
\arguments{
\item{mrobj}{A methylRaw object as returned by methylKit::unite()}

\item{destrand}{methylKit::unite() parameter; default: FALSE.
destrand=TRUE combines CpG methylation calls from both strands.}

\item{plot2file}{A logical determining whether plot output should be
sent to PDF files; default: FALSE.}
}
\value{
the data obtained from methylKit::unite()
}
\description{
cmpSamples()
This function will methylKit::unite() input samples and compare
them using methylKit::getCorrelation() and methylKit::PCASamples.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGhsm",
                       mincvrg=1,assembly="Amel-4.5")
  cmpSamples(AmHE,destrand=TRUE,plot2file=TRUE)

}
