% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpSites.R
\name{cmpSites}
\alias{cmpSites}
\title{cmpSites()
This function will}
\usage{
cmpSites(sample1hsm, sample1scd, sample1label, sample2hsm, sample2scd,
  sample2label, nbrpms)
}
\arguments{
\item{sample1hsm}{A methylRaw object}

\item{sample1scd}{A methylRaw object}

\item{sample1label}{A string}

\item{sample2hsm}{A methylRaw object}

\item{sample2scd}{A methylRaw object}

\item{sample2label}{A string}

\item{nbrpms}{integer}
}
\value{
a list of df's
}
\description{
cmpSites()
This function will
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  nbrpms  <- myfiles$parameters[myfiles$parameters$Variable == "TOTALNBRPMSITES",2]
  AmHEhsm <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                          type="CpGhsm", mincov=1,assembly="Amel-4.5")
  AmHEscd <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                          type="CpGscd", mincov=1,assembly="Amel-4.5")
  cmpSites(getData(AmHEhsm[[1]]),getData(AmHEscd[[1]]),"Am_HE_fr",
           getData(AmHEhsm[[2]]),getData(AmHEscd[[2]]),"Am_HE_rn",nbrpms)

}
