% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmStats.R
\name{cmStats}
\alias{cmStats}
\title{cmStats()
This function generates coverage and methylation statistics for the
  input data set.}
\usage{
cmStats(mrobj, covlist = c(10), outfile = "", plotfile = "")
}
\arguments{
\item{mrobj}{A methylRaw object}

\item{covlist}{a vector of coverage threshols; default: c(10)}

\item{outfile}{If specified then output is printed to the specified file.}

\item{plotfile}{If specified other than the default "", then plots
are saved in PDF file "plotfile".pdf; otherwise either no plots are
generated or the R default plot outputs are used.}
}
\value{
the coverage count for the highest threshold in covlist
}
\description{
cmStats()
This function generates coverage and methylation statistics for the
  input data set.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm", mincov=1,assembly="Amel-4.5")
  cmStats(AmHE[[1]],covlist=c(10,20),outfile="Am_HE-statistics.txt",plotfile="myplots")

}
