% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_dmsg.R
\name{det_dmsg}
\alias{det_dmsg}
\title{det_dmsg()
This function will get a list of genes with diff methylated CpG sites}
\usage{
det_dmsg(mrobj, genome_ann, threshold = 0.25, qvalue = 0.05, mc.cores = 1,
  outfile1 = "methyl_dmsites.txt", outfile2 = "methyl_dmgenes.txt")
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object.}

\item{genome_ann}{Genome annotation returned by get_genome_annotation()}

\item{threshold}{cutoff for percent methylation difference, default threshold = 0.25}

\item{qvalue}{cutoff for q-value, defualt q-value = 0.05}

\item{mc.cores}{Integer denoting how many cores should be used for parallel
diffential methylation calculations}

\item{outfile1}{File name to which diff sites are written}

\item{outfile2}{File name to which diff genes are written}
}
\value{
A Granges object that contains a list of genes that have diff methylated C sites
}
\description{
det_dmsg()
This function will get a list of genes with diff methylated CpG sites
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGhsm",
                       mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  meth_diff <- det_dmsg(AmHE,genome_ann,threshold=0.25,qvalue=0.05,mc.cores=4,
                        outfile1="AmHE-dmsites.txt", 
                        outfile2="AmHE-dmgenes.txt") 

}
