% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_dmt.R
\name{det_dmt}
\alias{det_dmt}
\title{det_dmt()
  This function generates list of differentially methylated CpG tiles and the
  corresponding genes, as determined by application of methylKit functions.}
\usage{
det_dmt(
  mrobj,
  genome_ann,
  wsize = 1000,
  stepsize = 1000,
  threshold = 25,
  qvalue = 0.01,
  mc.cores = 1,
  destrand = FALSE,
  outfile1 = "dmt.txt",
  outfile2 = "dmg.txt"
)
}
\arguments{
\item{mrobj}{A methylKit methylRaw or methylRawList object}

\item{genome_ann}{Genome annotation returned by get_genome_annotation()}

\item{wsize}{The window size for the sliding windows (tiles), default wsize = 1000}

\item{stepsize}{The step size for the sliding window starts, default stepsize = 1000}

\item{threshold}{Cutoff for percent methylation difference, default threshold = 25.0}

\item{qvalue}{Cutoff for q-value, default q-value = 0.01}

\item{mc.cores}{Integer denoting how many cores should be used for parallel
diffential methylation calculations}

\item{destrand}{methylKit::unite() parameter; default: FALSE.
destrand=TRUE combines CpG methylation calls from both strands.}

\item{outfile1}{File name to which differentially methylated tiles are written}

\item{outfile2}{File name to which differentially methylated genes are written}
}
\value{
A Granges object that contains a list of genes that have diff methylated C tiles
}
\description{
det_dmt()
  This function generates list of differentially methylated CpG tiles and the
  corresponding genes, as determined by application of methylKit functions.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGscd",
                       mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  mTlist <- det_dmt(AmHE,genome_ann,threshold=25.0,qvalue=0.01,mc.cores=4,
                    outfile1="AmHE-dmtiles.txt", 
                    outfile2="AmHE-dmgenes.txt") 

}
