% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_mprr.R
\name{det_mprr}
\alias{det_mprr}
\title{det_mprr()
This function generates coverage and methylation statistics for the
  input data set.}
\usage{
det_mprr(mrobj, ddset = c(1, 5), nr2d = 10L, outfstem = "",
  plotfile = "")
}
\arguments{
\item{mrobj}{A methylRaw object}

\item{outfstem}{If specified then output is printed to the specified file.}

\item{plotfile}{If specified other than the default "", then plots
are saved in PDF file "plotfile".pdf; otherwise either no plots are
generated or the R default plot outputs are used.}

\item{covlist}{a vector of coverage threshols; default: c(10)}
}
\value{
the coverage count for the highest threshold in covlist
}
\description{
det_mprr()
This function generates coverage and methylation statistics for the
  input data set.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm", mincov=1,assembly="Amel-4.5")
  det_mprr(AmHE[[1]],ddset=c(1,5),nr2d=10L,outfstem="Am_HE-regions",plotfile="myplots")

}
