% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_annotation.R
\name{get_genome_annotation}
\alias{get_genome_annotation}
\title{get_genome_annotation()
This function derives GRanges objects from the GFF3-formatted input file data.
The GFF3 input files should be specified by setup_BWASPR().}
\usage{
get_genome_annotation(inputdf)
}
\arguments{
\item{inputdf}{A data frame as returned by setup_BWASPR from the input data file.
The data frame contains the directories to specific generic features of DNA and the UTRflag.}
}
\value{
A list consisting of GRanges objects that describe generic features of DNA including:
  gene,exon,pcexon,promoter,CDS,fiveprimeUTR,threeprimeUTR,
  fiveprimeUTRunique,threeprimeUTRnotCDS,threeprimeUTRunique,ncexon
}
\description{
get_genome_annotation()
This function derives GRanges objects from the GFF3-formatted input file data.
The GFF3 input files should be specified by setup_BWASPR().
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  genome <- get_genome_annotation(myfiles$parameters)

}
