% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mdiff_info.R
\name{get_mdiff_info}
\alias{get_mdiff_info}
\title{get_mdiff_info()
This function will get a GRanges object contains differentially methylated sites between treatments
and get a GRanges object contains genes with differentially methylated sites}
\usage{
get_mdiff_info(mrobj, genome, threshold = 0.25, qvalue = 0.05,
  outfile1 = "methyl_diff_sites.txt", outfile2 = "methyl_diff_genes.txt")
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object.}

\item{genome}{Genome annotation returns by get_genome_annotation()}

\item{threshold}{cutoff for percent methylation difference, default threshold = 0.25}

\item{qvalue}{cutoff for q-value, defualt q-value = 0.05}
}
\value{
A Granges object that contains a list of genes that have diff methylated C sites
}
\description{
get_mdiff_info()
This function will get a GRanges object contains differentially methylated sites between treatments
and get a GRanges object contains genes with differentially methylated sites
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGhsm",
                       mincov=1,assembly="Amel-4.5")
  genome <- get_genome_annotation(myfiles$parameters)
  meth_diff <- get_mdiff_info(AmHE, genome)

}
