% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_dmgenes.R
\name{heatmap_dmgenes}
\alias{heatmap_dmgenes}
\title{heatmap_dmgenes()
This function will plot out the mC level pattern in genes that were identified with diff methylated C sites}
\usage{
heatmap_dmgenes(mrobj, mdiff, outfile = "hm_dmgenes.pdf")
}
\arguments{
\item{mrobj}{a methylRawList object}

\item{outfile}{dir/file name of the output figure}

\item{diff_genes}{diff genes}
}
\value{
A Granges object that contains a list of genes that have diff methylated C sites
}
\description{
heatmap_dmgenes()
This function will plot out the mC level pattern in genes that were identified with diff methylated C sites
}
\examples{
  mydatf  <- system.file("extdata","Am.dat",package="BWASPR")
  myparf  <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE    <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGhsm",
                          mincov=1,assembly="Amel-4.5")
  genome  <- get_genome_annotation(myfiles$parameters)
  mdiff   <- det_dmsg(AmHE, genome)
  heatmap_dmgenes(AmHE, meth_diff)

}
