% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_mrpr.R
\name{map_mrpr}
\alias{map_mrpr}
\title{map_mrpr()
This function maps the scd and hsm sites onto different genomic features
  (e.g. genes, exon, promoter, etc.)}
\usage{
map_mrpr(hsmrL, species, slabel, genome_ann, gnmsize, UTRflag, outfile)
}
\arguments{
\item{hsmrL}{List of data frames with hsm rich and poor regions, as put out by}

\item{species}{Label for the species being analyzed}

\item{slabel}{Label for the sample}

\item{genome_ann}{A list of GRanges objects that contains genome annotation.}

\item{gnmsize}{Genome size}

\item{UTRflag}{Numerical, indicating whether or not the annotation included UTRs (1 or 0)}

\item{outfile}{If specified, then the result is saved in the specified file name.}
}
\value{
Data frames describing genes overlapping with methylation-rich and poor regions
}
\description{
map_mrpr()
This function maps the scd and hsm sites onto different genomic features
  (e.g. genes, exon, promoter, etc.)
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  infiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  asmblv <- infiles$parameters[infiles$parameters$Variable == "ASSEMBLYVERSION",2]
  gnmsize <- as.numeric(infiles$parameters[infiles$parameters$Variable == "GENOMESIZE",2])
  UTRflag <- as.numeric(infiles$parameters[infiles$parameters$Variable == "UTRFLAGSET",2])
  AmHE <- mcalls2mkobj(infiles$datafiles,species="Am",study="HE",
                       sample=list("forager","nurse"),replicate=c(0),
                       type="CpGhsm",mincov=1,assembly=asmblv
                      )
  hsmrL <- det_mrpr(AmHE[[1]],"Am_HE_fr",ddset=c(1,5),outfile="dst-Am_HE_fr.txt",nr2d=10L,doplots=TRUE)
  ginfo <- get_genome_annotation(infiles$parameters)
  map_mrpr(hsmrL,species="Am",slabel="forager_hsm",ginfo,gnmsize,UTRflag=UTRflag,
           outfile="AmHE-region.map.txt"
          )

}
