% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcalls2mkobj.R
\name{mcalls2mkobj}
\alias{mcalls2mkobj}
\title{Setting up a BWASPR analysis structure}
\usage{
mcalls2mkobj(inputdf, species, study, type, assembly)
}
\arguments{
\item{datafile}{mcalls2mkobj()}
}
\value{
a methylBase object
}
\description{
Setting up a BWASPR analysis structure
}
