% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_annotation.R
\name{plot_genome_annotation}
\alias{plot_genome_annotation}
\title{plot_genome_annotation()
This function plots a representative figure of genome annotation}
\usage{
plot_genome_annotation(mrobj, genome_annotation_info, if_plot = FALSE,
  output_file_dir = "./BWASPR/R/genome_annotation.pdf",
  number_of_sites = 200)
}
\arguments{
\item{mrobj}{}

\item{genome_annotation}{}
}
\value{
A data frame that store the genome annotation data
}
\description{
plot_genome_annotation()
This function plots a representative figure of genome annotation
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  genome <- read_genome_info(myfiles$parameters)
  genome_annotation_info <- genome_annotation(AmHE, genome, save_output_data = FALSE)
  plot_genome_annotation(genome_annotation)


}
