% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_dmg.R
\name{rank_dmg}
\alias{rank_dmg}
\title{rank_dmg() 
This function sorts genes by different parameters and plots bar graph for demonstration}
\usage{
rank_dmg(dmsg.txt)
}
\arguments{
\item{dmsg.txt}{a tab delimited file generated by explore_dmsg()}
}
\value{
A list of data frames summarizing genes sorted by different parameter
}
\description{
rank_dmg() 
This function sorts genes by different parameters and plots bar graph for demonstration
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  dmsgList <- det_dmsg(AmHE,genome_ann,
                       threshold=25.0,qvalue=0.01,mc.cores=4,
                       destrand=TRUE,
                       outfile1="AmHE-dmsites.txt", 
                       outfile2="AmHE-dmgenes.txt")
  dmgprp <- show_dmsg(AmHE,dmsgList,min.nsites=10,destrand=TRUE,
                      outflabel="Am_HE")
  summaries <- explore_dmsg(AmHE,genome_ann,dmgprp,withglink="NCBIgene",
                            outflabel="Am_HE")

}
