% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_rbm.R
\name{rank_rbm}
\alias{rank_rbm}
\title{rank_rbm() 
  This function subsets the provided MethylRaw objects by the specified GRange
  and returns a list of dataframes containing the GRange regions ranked by
  methylation site density.}
\usage{
rank_rbm(mrobj, region.gr, rlabel = "sir", withglink = "", outflabel = "")
}
\arguments{
\item{mrobj}{A methyRaw/methRawList object}

\item{region.gr}{A GRanges object (typically based on the relevant genome annotation)}

\item{rlabel}{A string to identify the type of region analyzed}

\item{withglink}{Either NCBIgene or "" to indicate inculsion of a gene link column in the output}

\item{outflabel}{A string to identify the output dataframe}
}
\value{
A list of dataframes containing the GRange regions ranked by methylation
  site density
}
\description{
rank_rbm() 
  This function subsets the provided MethylRaw objects by the specified GRange
  and returns a list of dataframes containing the GRange regions ranked by
  methylation site density.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  summaries <- rank_rbm(AmHE,region.gr=genome_ann$gene,rlabel="sir",
                        withglink="NCBIgene",outflabel="Am_HE_gene")

}
