% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome_info.R
\name{read_genome_info}
\alias{read_genome_info}
\title{read_genome_info()
This function will read the genome information}
\usage{
read_genome_info(inputdf)
}
\arguments{
\item{inputdf}{A data frame as returned by setup_BWASPR from the input
data file.  Each row corresponds to a BWASP-generated output, with columns
indicating Species, Study, Sample, Replicate, Type, and File.}
}
\value{
A list consisting of genome information including
gene.gr,
exon.gr,
pcexon.gr,
promoter.gr,
CDS.gr,
fiveprimeUTR.gr,
threeprimeUTR.gr,
fiveprimeUTR_unique.gr,
threeprimeUTRnotCDS.gr,
threeprimeUTR_unique.gr,
ncexon.grl
}
\description{
read_genome_info()
This function will read the genome information
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  genome <- read_genome_info(myfiles$parameters)

}
