% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_BWASP.R
\name{setup_BWASPR}
\alias{setup_BWASPR}
\title{Setting up a BWASPR analysis structure}
\usage{
setup_BWASPR(datafile, parfile)
}
\arguments{
\item{datafile, }{parfile
setup_BWASPR()}
}
\value{
a list consisting of a data frame with the data input files) and a 
        data frame with parameters
}
\description{
Setting up a BWASPR analysis structure
}
