% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dmsg.R
\name{show_dmsg}
\alias{show_dmsg}
\title{show_dmsg() 
  This function generates will plot heatmaps of methylation level of sites in dmgenes}
\usage{
show_dmsg(mrobj, dmsg, destrand = FALSE, min.nsites = 2,
  max.nsites = 60, min.pdmsites = 10, mc.cores = 1, outflabel = "")
}
\arguments{
\item{mrobj}{A methyRaw object or a methyRawList object}

\item{dmsg}{A list containing GRanges objects dmsites and dmgenes returned by
det_dmsg()}

\item{destrand}{methylKit::unite() parameter; default: FALSE.
destrand=TRUE combines CpG methylation calls from both strands}

\item{min.nsites}{Required minimal number of msites in a gene for the gene
to be displayed in the heatmap pdf}

\item{max.nsites}{Required maximal number of msites in a gene for the gene
to be displayed in the heatmap pdf}

\item{min.pdmsites}{Required minimal percent of dmsites among the msites in
a gene for the gene to be displayed in the heatmap pdf}

\item{mc.cores}{Integer denoting how many cores should be used for parallel
diffential methylation calculations}

\item{outflabel}{A string to identify the study in the output file}
}
\value{
A data frame
}
\description{
show_dmsg() 
  This function generates will plot heatmaps of methylation level of sites in dmgenes
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  samplelist <- list("forager","nurse")
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       sample=samplelist,replicate=c(0),
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  dmsgList <- det_dmsg(AmHE,genome_ann,
                       threshold=25.0,qvalue=0.01,mc.cores=4,destrand=TRUE,
                       outfile1="AmHE-dmsites.txt", 
                       outfile2="AmHE-dmgenes.txt")
  dmgprp <- show_dmsg(AmHE,dmsgList,destrand=TRUE,
                      min.nsites=2,max.nsites=60,min.pdmsites=10,
                      mc.cores=4,outflabel="Am_HE")

}
