% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_mrobj.R
\name{subset_mrobj}
\alias{subset_mrobj}
\title{subset_mrobj() 
This function subsets the mrobj by the GRanges that the user provides 
and returns a List of dataframes containing the msites info summries and
save the dataframes as tab delimited files}
\usage{
subset_mrobj(mrobj, region.gr, outflabel = "")
}
\arguments{
\item{mrobj}{A methyRaw/methRawList object or a methyRawList object}

\item{region.gr}{A Granges object that the user provieds}

\item{outflabel}{A string to identify the output dataframe}
}
\value{
A list of data frames that contains the msites info within the GRanges the user provided
}
\description{
subset_mrobj() 
This function subsets the mrobj by the GRanges that the user provides 
and returns a List of dataframes containing the msites info summries and
save the dataframes as tab delimited files
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  promoter_mrobj <- subset_mrobj(AmHE,region.gr=genome_ann$promoter,
                                 outflabel="Am_HE_promoter")

}
