% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_methylome.R
\name{annotate_methylome}
\alias{annotate_methylome}
\title{annotate_methylome()
This function annotates the methylome with different genomic features (e.g. genes, exon, promoter, etc.)}
\usage{
annotate_methylome(
  mrobj,
  genome_ann,
  destrand = FALSE,
  mc.cores = mc.cores,
  outfile = "methylome_ann.txt"
)
}
\arguments{
\item{mrobj}{A methylRawList object that is generated by mcalls2mkobj()}

\item{genome_ann}{A list of GRanges objects that contain genome annotation.}

\item{destrand}{methylKit::unite() parameter; default: FALSE.
destrand=TRUE combines CpG methylation calls from both strands}

\item{mc.cores}{Integer denoting how many cores should be used for parallel
diffential methylation calculations}

\item{outfile}{If specified as other than "", the result is saved in the
specified file.}
}
\value{
A table that displays site methylation percentages and annotation
}
\description{
annotate_methylome()
This function annotates the methylome with different genomic features (e.g. genes, exon, promoter, etc.)
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  genome_ann <- get_genome_annotation(myfiles$parameters)
  annotate_methylome(AmHE,genome_ann,mc.cores=4,
                     outfile="AmHE_methylome_ann.txt")

}
