% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmpSamples.R
\name{cmpSamples}
\alias{cmpSamples}
\title{cmpSamples()
This function will methylKit::unite() input samples and compare
them using methylKit::getCorrelation() and methylKit::PCASamples.}
\usage{
cmpSamples(mrobj, destrand = FALSE, mc.cores = 1, plotfile = "")
}
\arguments{
\item{mrobj}{A methylRaw object as returned by methylKit::unite()}

\item{destrand}{methylKit::unite() parameter; default: FALSE.
destrand=TRUE combines CpG methylation calls from both strands.}

\item{mc.cores}{Integer denoting how many cores should be used for parallel
diffential methylation calculations}

\item{plotfile}{If specified other than the default "", then plots
are saved in PDF file "plotfile".pdf; otherwise either no plots are
generated or the R default plot outputs are used.}
}
\value{
the data obtained from methylKit::unite()
}
\description{
cmpSamples()
This function will methylKit::unite() input samples and compare
them using methylKit::getCorrelation() and methylKit::PCASamples.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       sample="all",replicate=c(0),
                       type="CpGhsm", mincov=1,
                       assembly="Amel-4.5")
  cmpSamples(AmHE,destrand=TRUE,mc.cores=4,plotfile="myplots")

}
