% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmStats.R
\name{cmStats}
\alias{cmStats}
\title{cmStats()
This function generates coverage and methylation statistics for the
  input data set.}
\usage{
cmStats(mrobj, type = "CpGhsm", covlist = c(10), locount = 100,
  hicount = 1000, outfile = "", plotfile = "")
}
\arguments{
\item{mrobj}{A methylRaw object}

\item{type}{The type should be CpGhsm or CpGscd, as per BWASP output.}

\item{covlist}{A vector of coverage threshols; default: c(10)}

\item{locount}{Low coverage threshold for statistics on a range of sites; default: c(100)}

\item{hicount}{High coverage threshold for statistics on a range of sites; default: c(1000)}

\item{outfile}{If specified then output is printed to the specified file.}

\item{plotfile}{If specified other than the default "", then plots
are saved in PDF file "plotfile".pdf; otherwise either no plots are
generated or the R default plot outputs are used.}
}
\value{
the coverage count for the highest threshold in covlist
}
\description{
cmStats()
This function generates coverage and methylation statistics for the
  input data set.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm", mincov=1,assembly="Amel-4.5")
  cmStats(AmHE[[1]],type="CpGhsm",covlist=c(10,20),locount=100,hicount=1000,
          outfile="Am_HE-statistics.txt",plotfile="myplots")

}
