% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmStats.R
\name{cmStats}
\alias{cmStats}
\title{cmStats()
This function will}
\usage{
cmStats(mrobj, covlist = c(10), plotfile = "not_set")
}
\arguments{
\item{mrobj}{A methylRaw object}

\item{covlist}{a vector of coverage threshols; default: c(10)}

\item{plotfile}{If specified other than the default "not_set", then plots
are saved in PDF file "plotfile".pdf; otherwise either no plots are
generated or the R default plot outputs are used.}
}
\value{
the coverage count for the highest threshold in covlist
}
\description{
cmStats()
This function will
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm", mincov=1,assembly="Amel-4.5")
  cmStats(AmHE,plotfile="myplots")

}
