% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/det_dmsg.R
\name{det_dmsg}
\alias{det_dmsg}
\title{det_dmsg()
  This function generates lists of differentially methylated CpG sites and the
  corresponding genes, as determined by application of methylKit functions.}
\usage{
det_dmsg(mrobj, genome_ann, threshold = 25, qvalue = 0.01,
  mc.cores = 1, destrand = FALSE, outfile1 = "methyl_dmsites.txt",
  outfile2 = "methyl_dmgenes.txt")
}
\arguments{
\item{mrobj}{A methylKit methylRaw or methylRawList object}

\item{genome_ann}{Genome annotation returned by get_genome_annotation()}

\item{threshold}{Cutoff for percent methylation difference, default threshold = 25.0}

\item{qvalue}{Cutoff for q-value, default q-value = 0.01}

\item{mc.cores}{Integer denoting how many cores should be used for parallel
diffential methylation calculations}

\item{destrand}{methylKit::unite() parameter; default: FALSE.
destrand=TRUE combines CpG methylation calls from both strands.}

\item{outfile1}{File name to which diff sites are written}

\item{outfile2}{File name to which diff genes are written}
}
\value{
A list of GRanges objects that contain data on differentially methylated
  sites and genes with such sites
}
\description{
det_dmsg()
  This function generates lists of differentially methylated CpG sites and the
  corresponding genes, as determined by application of methylKit functions.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGhsm",
                       mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  meth_diff <- det_dmsg(AmHE,genome_ann,threshold=25.0,qvalue=0.01,mc.cores=4,
                        outfile1="AmHE-dmsites.txt", 
                        outfile2="AmHE-dmgenes.txt") 

}
