% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_dmsg.R
\name{explore_dmsg}
\alias{explore_dmsg}
\title{explore_dmsg() 
This function sorts genes by meth_level per 10kb for individual samples, 
and sorts genes by meth_diff per 10kb for each comparison}
\usage{
explore_dmsg(mrobj, genome_ann, show_dmsg)
}
\arguments{
\item{mrobj}{A methyRaw/methRawList object or a methyRawList object}

\item{genome_ann}{Genome annotation returned by get_genome_annotation()}

\item{show_dmsg}{a list of dataframe contains dmsg generated by show_dmsg()}
}
\value{
MISSING
}
\description{
explore_dmsg() 
This function sorts genes by meth_level per 10kb for individual samples, 
and sorts genes by meth_diff per 10kb for each comparison
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  meth_diff <- det_dmsg(AmHE,genome_ann,
                        threshold=25.0,qvalue=0.01,mc.cores=4,
                        outfile1="AmHE-dmsites.txt", 
                        outfile2="AmHE-dmgenes.txt")
  show_dmsg <- show_dmsg(AmHE,meth_diff)
  explore_dmsg(AmHE,genome_ann,show_dmsg)

}
