% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_dmsg.R
\name{explore_dmsg}
\alias{explore_dmsg}
\title{explore_dmsg() 
This function sorts genes by meth_level per 10kb for individual samples
and sorts genes by meth_diff per 10kb for each comparison; sample
percent methylation per site values are compared with the Wilcoxon
signed rank test}
\usage{
explore_dmsg(mrobj, genome_ann, dmgprp, maxgwidth, minnbrdmsites,
  withglink = "NCBIgene", outflabel = "")
}
\arguments{
\item{mrobj}{A methyRaw/methRawList object or a methyRawList object}

\item{genome_ann}{Genome annotation returned by get_genome_annotation()}

\item{dmgprp}{A list of data frames containing dmsg data generated by
show_dmsg()}

\item{maxgwidth}{An integer specifying the maximum gene width for a gene
to be listed in the output tables [default: -1, interpreted to mean
no restriction]}

\item{minnbrdmsites}{An integer specifying the minimum number of
differentially methylated sites for a gene to be listed in the pairwise
sample comparisons [default: 1]}

\item{withglink}{Either NCBIgene or "" to indicate inclusion of a
gene link column in the output}

\item{outflabel}{A string to identify the study in the output file}
}
\value{
A list of data frames summarizing single sample and pairwise
  comparison gene properties
}
\description{
explore_dmsg() 
This function sorts genes by meth_level per 10kb for individual samples
and sorts genes by meth_diff per 10kb for each comparison; sample
percent methylation per site values are compared with the Wilcoxon
signed rank test
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  dmsgList <- det_dmsg(AmHE,genome_ann,
                       threshold=25.0,qvalue=0.01,mc.cores=4,
                       destrand=TRUE,
                       outfile1="AmHE-dmsites.txt", 
                       outfile2="AmHE-dmgenes.txt")
  dmgprp <- show_dmsg(AmHE,dmsgList,min.nsites=10,destrand=TRUE,
                      outflabel="Am_HE")
  summaries <- explore_dmsg(AmHE,genome_ann,dmgprp,maxgwidth=20000,
                            minnbrdmsites=2, withglink="NCBIgene",
                            outflabel="Am_HE")

}
