% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_annotation.R
\name{genome_annotation}
\alias{genome_annotation}
\title{genome_annotation()
This function annotates the methylated sites with different genomic features (e.g. genes, exon, promoter, etc.)}
\usage{
genome_annotation(mrobj, genome_info, save_output_data = TRUE,
  outputdata = "genome_annotation_data.Rda")
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object.}

\item{genome_info}{A list of GRanges objects that contains genome infomation.}

\item{save_output_data}{If specified other than the default "TRUE", then results
are saved in Rda file that is specified by @param outputdata; otherwise either no Rda file is generated.}

\item{outputdata}{If specified, then the result is saved in the specified file name; otherwise the result
is saved in '$wd/genome_annotation_data.Rda'}
}
\value{
A table ?that stores the methylation information and genome_annotation
}
\description{
genome_annotation()
This function annotates the methylated sites with different genomic features (e.g. genes, exon, promoter, etc.)
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  genome <- read_genome_info(myfiles$parameters)
  genome_annotation_info <- genome_annotation(AmHE, genome, save_output_data = FALSE)
  genome_annotation(AmHE, genome, outputdata = 'AmHE_genome_annotation.Rda')

}
