% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff_genes.R
\name{get_diff_genes}
\alias{get_diff_genes}
\title{get_diff_genes()
This function will get a list of genes with diff methylated CpG sites}
\usage{
get_diff_genes(mrobj, genome, threshold = 0, qvalue = 1, mc.cores = 1)
}
\arguments{
\item{mrobj}{a methylRawList object}

\item{genome}{output from read_genome_info}

\item{threshold}{Threshold of methylation level change}

\item{qvalue}{criterion for determination of significance}

\item{mc.cores}{Number of cores used by methylKit functions}
}
\value{
A Granges object that contains a list of genes that have diff methylated C sites
}
\description{
get_diff_genes()
This function will get a list of genes with diff methylated CpG sites
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGhsm",
                       mincvrg=1,assembly="Amel-4.5")
  genome <- read_genome_info(myfiles$parameters)
  meth_diff <- get_diff_genes(AmHE, genome)

}
