% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diff_genes.R
\name{get_diff_genes}
\alias{get_diff_genes}
\title{get_diff_genes()
??This function will get a list of genes with diff methylated CpG sites}
\usage{
get_diff_genes(mrobj, genome, threshold = 0.25, qvalue = 0.05)
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object.}

\item{threshold}{cutoff for percent methylation difference, default threshold = 0.25}

\item{qvalue}{cutoff for q-value, defualt q-value = 0.05}
}
\value{
A Granges object that contains a list of genes that have diff methylated C sites
}
\description{
get_diff_genes()
??This function will get a list of genes with diff methylated CpG sites
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",type="CpGhsm",
                       mincvrg=1,assembly="Amel-4.5")
  genome <- read_genome_info(myfiles$parameters)
  meth_diff <- get_diff_genes(AmHE, genome)

}
