% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_methylome.R
\name{map_methylome}
\alias{map_methylome}
\title{map_methylome()
This function annotates maps the scd and hsm site onto different genomic
  features (e.g. genes, exon, promoter, etc.)}
\usage{
map_methylome(studymk, slabel, studymc, clabel, genome_info, gnmsize, UTRflag,
  outputdata = "some.Rda")
}
\arguments{
\item{studymk}{methylRaw object representing an hsm sample}

\item{slabel}{Label for the sample}

\item{studymc}{methylRaw object representing the scd control}

\item{clabel}{Label for the control sample}

\item{genome_info}{A list of GRanges objects that contains genome infomation.}

\item{gnmsize}{Genome size}

\item{UTRflag}{Numerical, indicating whether or not the annotation included UTRs (1 or 0)}

\item{outputdata}{If specified, then the result is saved in the specified file name; otherwise the result
is saved in '$wd/something.Rda'}
}
\value{
Should return something useful
}
\description{
map_methylome()
This function annotates maps the scd and hsm site onto different genomic
  features (e.g. genes, exon, promoter, etc.)
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  gnmsize <- as.numeric(infiles$parameters[infiles$parameters$Variable == "GENOMESIZE",2])
  UTRflag <- as.numeric(infiles$parameters[infiles$parameters$Variable == "UTRFLAGSET",2])
  ginfo <- get_genome_annotation(myfiles$parameters)
  map_methylome(AmHE, ginfo, gnmsize, outputdata='AmHE_map_methylome.Rda')

}
