% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_plot_diff_genes_pattern.R
\name{plot_diff_genes_pattern}
\alias{plot_diff_genes_pattern}
\title{plot_diff_genes_pattern()
This function will plot out the mC level pattern in genes that were identified with diff methylated C sites}
\usage{
plot_diff_genes_pattern(mrobj, diff_genes, output_figure)
}
\arguments{
\item{diff_genes}{diff genes}

\item{output_figure}{name of the output figure}

\item{mcobj}{a methylRawList object}
}
\value{
A Granges object that contains a list of genes that have diff methylated C sites
}
\description{
plot_diff_genes_pattern()
This function will plot out the mC level pattern in genes that were identified with diff methylated C sites
}
\examples{
  meth_diff_genes <- get_diff_genes(mcobj)

}
