% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_annotation.R
\name{plot_genome_annotation}
\alias{plot_genome_annotation}
\title{plot_genome_annotation()
This function plots a representative figure of genome annotation}
\usage{
plot_genome_annotation(mrobj, genome_annotation_info, plot = FALSE,
  output_file = "genome_annotation.pdf", number_of_sites = 200)
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object.}

\item{plot}{TRUE|FALSE. If TRUE, plot the heatmap figure, if FALSE, save the heatmap figure.}

\item{output_file}{?An option to specify the dir/file names for output figure.}

\item{number_of_sites}{Number of methylated sites that are plotted on figure}

\item{genome_annotation}{A table generated by genome_annotation()
?that stores the methylation information and genome_annotation}
}
\value{
A data frame that store the genome annotation data
}
\description{
plot_genome_annotation()
This function plots a representative figure of genome annotation
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  genome <- read_genome_info(myfiles$parameters)
  genome_annotation_info <- genome_annotation(AmHE, genome, save_output_data = FALSE)
  plot_genome_annotation(AmHE, genome_annotation_info)

}
