% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_annotation.R
\name{plot_genome_annotation}
\alias{plot_genome_annotation}
\title{plot_genome_annotation()
This function plots a heatmap to show the annotation of the first 200 methylated sites}
\usage{
plot_genome_annotation(mrobj, methylome_annotation, plot = FALSE,
  output_file = "methylome_annotation.pdf")
}
\arguments{
\item{mrobj}{A methylRaw object or a methylRawList object.}

\item{methylome_annotation}{A table generated by annotate_methylome()
that stores the methylome information and methylome annotation information}

\item{plot}{TRUE|FALSE. If TRUE, plot the heatmap figure, if FALSE, save the heatmap figure.}

\item{output_file}{An option to specify the dir/file names for output figure.}
}
\value{
A data frame that store the genome annotation data
}
\description{
plot_genome_annotation()
This function plots a heatmap to show the annotation of the first 200 methylated sites
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles)
  genome_ann <- get_genome_annotation(myfiles$parameters)
  methylome_annotation <- annotate_methylome(AmHE,genome_ann,outfile="")
  plot_genome_annotation(AmHE,methylome_annotation)

}
