% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_rbm.R
\name{rank_rbm}
\alias{rank_rbm}
\title{rank_rbm() 
This function subsets the mrobj by the GRanges that the user provides 
and returns a List of dataframes containing the msites info summries and
save the dataframes as tab delimited files}
\usage{
rank_rbm(mrobj, region.gr, rlabel = "sir", withglink = "", outflabel = "")
}
\arguments{
\item{mrobj}{A methyRaw/methRawList object or a methyRawList object}

\item{region.gr}{A Granges object that the user provieds}

\item{rlabel}{A string to identify the type of region analyzed}

\item{withglink}{Either NCBIgene or "" to indicate inculsion of a gene link column in the output}

\item{outflabel}{A string to identify the output dataframe}
}
\value{
A list of data frames that contains the msites info within the GRanges the user provided
}
\description{
rank_rbm() 
This function subsets the mrobj by the GRanges that the user provides 
and returns a List of dataframes containing the msites info summries and
save the dataframes as tab delimited files
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  summaries <- rank_rbm(AmHE,region.gr=genome_ann$gene,rlabel="sir",
                       withglink="NCBIgene",outflabel="Am_HE_gene")

}
