% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome_info.R
\name{read_genome_info}
\alias{read_genome_info}
\title{read_genome_info()
This function will read the gff3 (generic feature format that describes gene and other features of DNA)
files according to specified selection criteria generated by setup_BWASPR()
and read the data into GRranges objects.}
\usage{
read_genome_info(inputdf)
}
\arguments{
\item{inputdf}{A data frame as returned by setup_BWASPR from the input data file.
The data frame contains the directories to specific generic features of DNA and the UTRflag.}
}
\value{
A list consisting of GRanges objects that describe generic features of DNA including:
  gene,exon,pcexon,promoter,CDS,fiveprimeUTR,threeprimeUTR,
  fiveprimeUTR_unique,threeprimeUTRnotCDS,threeprimeUTR_unique,ncexon.
}
\description{
read_genome_info()
This function will read the gff3 (generic feature format that describes gene and other features of DNA)
files according to specified selection criteria generated by setup_BWASPR()
and read the data into GRranges objects.
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  genome <- read_genome_info(myfiles$parameters)

}
