% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genome_info_genomation.R
\name{read_genome_info_genomation}
\alias{read_genome_info_genomation}
\title{read_genome_info_genomation()
This function converts the gff3 formated data.frame into a list with GRanges objects.
The gff3 files directories were specified by setup_BWASPR()}
\usage{
read_genome_info_genomation(inputdf)
}
\arguments{
\item{inputdf}{A data frame as returned by setup_BWASPR from the input data file.
The data frame contains the directories to specific generic features of DNA and the UTRflag.}
}
\value{
A list consisting of GRanges objects that describe generic features of DNA including:
  gene,exon,pcexon,promoter,CDS,fiveprimeUTR,threeprimeUTR,
  fiveprimeUTR_unique,threeprimeUTRnotCDS,threeprimeUTR_unique,ncexon
}
\description{
read_genome_info_genomation()
This function converts the gff3 formated data.frame into a list with GRanges objects.
The gff3 files directories were specified by setup_BWASPR()
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  genome <- read_genome_info_genomation(myfiles$parameters)

}
