% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_BWASP.R
\name{setup_BWASPR}
\alias{setup_BWASPR}
\title{setup_BWASPR()
Setting up a BWASPR analysis structure}
\usage{
setup_BWASPR(datafile, parfile)
}
\arguments{
\item{datafile}{A BWASP data description file consisting of rows of
tab-deliminted entries with consecutive labesl for Species, Study,
Sample, Replicate, Type, and BWASP mcalls file location.}

\item{parfile}{A BWASPR parameter file with parameter settings used in
in the analysis.}
}
\value{
A list consisting of a data frame with the data input files and a 
        data frame with parameters
}
\description{
setup_BWASPR()
Setting up a BWASPR analysis structure
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)

}
