% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dmsg.R
\name{show_dmsg}
\alias{show_dmsg}
\title{show_dmsg() 
This function will plot heatmap of methylation level of isites in dmgenes}
\usage{
show_dmsg(mrobj, dmgenes.file, min.nsites = 2, query.genes = "")
}
\arguments{
\item{mrobj}{A methyRaw object or a methyRawList object}

\item{dmgenes.file}{A tab delimited file stores genes with differential methylated sites}

\item{query.genes}{A list of gene IDs}

\item{min.sites}{minimal number of msites per genes}
}
\value{
MISSING - PROBABLY SHOULD RETURN data frame
}
\description{
show_dmsg() 
This function will plot heatmap of methylation level of isites in dmgenes
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  meth_diff <- det_dmsg(AmHE,genome_ann,
                        threshold=0.25,qvalue=0.25,mc.cores=4,
                        outfile1="AmHE-dmsites.txt", 
                        outfile2="AmHE-dmgenes.txt")
  show_dmsg(AmHE,'AmHE-dmgenes.txt')

}
