% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dmsg.R
\name{show_dmsg}
\alias{show_dmsg}
\title{show_dmsg() 
  This function generates will plot heatmaps of methylation level of sites in dmgenes}
\usage{
show_dmsg(mrobj, dmsg, min.nsites = 2, destrand = TRUE)
}
\arguments{
\item{mrobj}{A methyRaw object or a methyRawList object}

\item{dmsg}{A list containing GRanges objects dmsites and dmgenes returned by
det_dmsg()}

\item{min.nsites}{Minimal number of msites per genes}

\item{destrand}{methylKit::unite() parameter; default: FALSE.
destrand=TRUE combines CpG methylation calls from both strands}
}
\value{
A data frame
}
\description{
show_dmsg() 
  This function generates will plot heatmaps of methylation level of sites in dmgenes
}
\examples{
  mydatf <- system.file("extdata","Am.dat",package="BWASPR")
  myparf <- system.file("extdata","Am.par",package="BWASPR")
  myfiles <- setup_BWASPR(datafile=mydatf,parfile=myparf)
  AmHE <- mcalls2mkobj(myfiles$datafiles,species="Am",study="HE",
                       type="CpGhsm",mincov=1,assembly="Amel-4.5")
  genome_ann <- get_genome_annotation(myfiles$parameters)
  meth_diff <- det_dmsg(AmHE,genome_ann,
                        threshold=25.0,qvalue=0.01,mc.cores=4,
                        outfile1="AmHE-dmsites.txt", 
                        outfile2="AmHE-dmgenes.txt")
  show_dmsg <- show_dmsg(AmHE,meth_diff)

}
